Move unit test for chain module custom storage entitites
