+++ framework/test/mocha/unit/modules/chain/components/storage/entitites/peer.js
@@ -18,13 +18,13 @@
 const {
 	BaseEntity,
 	Peer,
-} = require('../../../../../../src/components/storage/entities');
-const storageSandbox = require('../../../../common/storage_sandbox');
-const peersFixtures = require('../../../../fixtures/peers');
+} = require('../../../../../../../../src/components/storage/entities');
+const storageSandbox = require('../../../../../../common/storage_sandbox');
+const peersFixtures = require('../../../../../../fixtures/peers');
 const {
 	NonSupportedFilterTypeError,
 	NonSupportedOptionError,
-} = require('../../../../../../src/components/storage/errors');
+} = require('../../../../../../../../src/components/storage/errors');
 
 // eslint-disable-next-line mocha/no-skipped-tests
 describe('Peer', () => {
